+++ core/src/trezor/ui/model/common/__init__.py
@@ -1,0 +1,9 @@
+"""
+The model/common module contains code that is used by both model/tt and model/t1.
+"""
+
+if False:
+    from typing import List, Union, Dict
+
+    TextContent = Union[str, int]
+    ListsArg = Dict[str, List[str]]
